Fix cross-platform release provenance validation
